Set up for splitting group tasks
